requirements-test.txt
apache-libcloud
azure-storage-blob >= 12.0, != 12.9.0  # version 12.9.0 breaks mypy https://github.com/Azure/azure-sdk-for-python/pull/20891
pytest
python-cephlibs
requests_mock[fixture] >= 1.9
requests_toolbelt
types-pyyaml
types-requests
